+++ src/client/src/AppTheme.ts
@@ -70,6 +70,7 @@ const themePalette: AppThemePalette = {
     filterItemActive: "#1C2834",
     listItemActive: "#DFE4E9",
     backdrop: "rgba(255,255,255,0.7)",
+    tabFocus: "rgba(100, 95, 228, 0.32)",
   },
   transparent: "rgba(0,0,0,0)",
   buttonStates: {
@@ -502,9 +503,13 @@ export const theme = createTheme({
       styleOverrides: {
         root: {
           fontFamily: "Inter",
-          fontWeight: "bold",
+          fontWeight: 400,
           textTransform: "none",
           fontSize: "16px",
+          "&.Mui-selected": {
+            color: themePalette.background.menuItemActive,
+            fontWeight: 500,
+          },
         },
       },
     },
@@ -516,6 +521,10 @@ export const theme = createTheme({
         },
         scroller: {
           overflow: "visible !important",
+        },
+        indicator: {
+          backgroundColor: themePalette.background.menuItemActive,
+          transition: "none",
         },
       },
     },