+++ src/client/src/components/styledTabComponents.tsx
@@ -47,6 +47,22 @@ export const BoreholeTab = styled(({ ...props }: BoreholeTabProps) => <Tab disab
     top: "1px",
   },
   "&.Mui-focusVisible": {
-    backgroundColor: "rgba(100, 95, 228, 0.32)",
+    backgroundColor: theme.palette.background.tabFocus,
+  },
+}));
+
+export const TabsWithDivider = styled(Tabs)(({ theme }) => ({
+  borderBottom: `1px solid ${theme.palette.border.light}`,
+  "& .MuiTabs-indicator": {
+    transform: "translateY(1px)",
+  },
+}));
+
+export const TabWithContent = styled(({ ...props }: BoreholeTabProps) => <Tab disableRipple {...props} />, {
+  shouldForwardProp: prop => prop !== "hasContent",
+})(({ theme, hasContent }) => ({
+  color: hasContent === false ? theme.palette.buttonStates.outlined.disabled.color : theme.palette.secondary.main,
+  "&.Mui-focusVisible": {
+    backgroundColor: theme.palette.background.tabFocus,
   },
 }));