+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Show loading indicator when importing data.
 - Select input fields now filter and display only the options that match the user's input.
 - Remove type from profile table.
+- Lithology, lithostratigraphy and chronostratigraphy are combined into one menu item.
 
 ### Fixed
 - Boreholes could not be exported in the _view_ environment.